Break tests out to properly namespace in directory
