+++ spec/requests/api/v1/merchant_request_spec.rb
@@ -91,34 +91,7 @@ RSpec.describe 'Merchants API', type: :request do
     expect(json[:message]).to eq("Couldn't find Merchant with 'id'=0")
   end
 
-  describe 'items merchant' do 
-    it "returns the merchant for an item" do
-      merchants = create_list(:merchant, 3)
-      merch_items = create_list(:item, 10, merchant: merchants.first)
-      item = merch_items.third 
-
-      get api_v1_item_merchant_index_path(item)
-
-      json = parse_json
-      json_data = parse_json[:data]
-      
-      expect(json).to be_a(Hash)
-      expect(response.status).to eq(200)
-      expect(json_data).to be_a(Hash)
-      expect(json_data[:id].to_i).to eq(merchants.first[:id])
-      expect(json_data[:attributes][:name]).to eq(merchants.first[:name])
-    end
+  it 'can search for a merchant' do 
     
-    it 'cannot find item' do 
-      merchants = create_list(:merchant, 3)
-      merch_items = create_list(:item, 10, merchant: merchants.first)
-
-      get api_v1_item_merchant_index_path(0)
-
-      json = parse_json
-      
-      expect(response.status).to eq(404)
-      expect(json[:message]).to eq("Couldn't find Item with 'id'=0")
-    end
   end
 end
